chore: fix cms build cache (database)
Here is an error I was getting locally:
  Unable to connect to database with message: SQLSTATE[HY000]: General error: 11 malformed database
  schema (block_content__body_bundle) - index block_content__body_bundle already exists.

After some debugging it was found:
- when cms `build` command is executed, it generates `.sqlite` file (no other db files exist)
- after some interactions with database, new files created: `.sqlite-shm` and `.sqlite-wal`
- when turbo restores cms build cache, it overrides `.sqlite` from cache, but leaves other db files
- this causes the malformed database error

To workaround this, we fire `select * from node` query at the end of the build. This forces the
creation of `.sqlite-shm` and `.sqlite-wal` files. So next time turbo restores the cache, all db
files are overridden, and there are no issues with the database integrity.