chore: make publisher database temporary
It had occurred a few times on live projects that the Publisher's database became corrupted. During
the subsequent deployment, the Publisher attempted to start but could not due to the broken
database.

We could have the Publisher recreate the database if it becomes corrupted. However, in reality, the
build history isn't truly valuable, so we only make it temporary.