chore: fix `composer install` in docker
After upgrading Turborepo to 2.0 we got this error in "@custom/cms:prep:composer":
  Composer plugins have been disabled for safety in this non-interactive session.
  Set COMPOSER_ALLOW_SUPERUSER=1 if you want to allow plugins to run as root/super user.
  Do not run Composer as root/super user! See https://getcomposer.org/root for details

From https://github.com/composer/composer/issues/11073 it looks like it is fine to run Composer as
superuser - Composer tries to detect if it's running in a Docker context and suppresses the warning.
In our case, the check fails for some reason. So we have to manually set COMPOSER_ALLOW_SUPERUSER.